requirements.txt
accelerate==1.6.0
datasets==3.5.0
huggingface-hub==0.30.2
joblib==1.4.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
notebook_shim==0.2.4
numpy==2.2.5
openai==1.75.0
pandas==2.2.3
python-dotenv==1.1.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
tokenizers==0.21.1
torch==2.6.0
transformers==4.51.3
uvicorn==0.34.2
xgboost==3.0.0
gradio==5.25.2